Made scaffold of project, as well as used generic gitignore. This is a trello comment I left but I'm pasting it here to in case:
"also, for the project scaffold, just a basic summary if you’re not aware with flutter structure heres a simple layout:

lib folder - this is the main app code, it’ll contain like most of our source code, and within it its gonna have a bunch of other folders for useful stuff. most important for us will be the sub folders for main.dart, screens, and services. main.dart is the entry point, screens is the diff pages, services is for our backend connection.

we have a web and windows folder, we only care about the web tbh, i accidentally initialized the flutter project with a windows project bc i forgot the console command, but we can ignore it because windows is like if we wanna make it a windows app, but web is the folder for our web platform files that we need to actually run our app on a web browser. we dont actually write much here, flutter handles it.

also, to actually test it on web, u just run in console flutter run -d chrome"